All Pieces move correctly except the En Passant move
Castling works (phew). Wrote the grounds from which en passant can be
coded to (do some turn handling but im tired rn :p) #1 done. Swapping a
pawn at the endline is still  a thing that needs to be bugfixed but
mostly done. Besides that, all pieces can move and only move the way
they are allowed to by the rules